Fixed casting tripworkaround, commenting unused method

+++ Quecklink/src/main/java/QueclinkProto/Gtigf.java
@@ -78,8 +78,8 @@ public class Gtigf extends QueclinkReport{
 		TripShutdown scopeEvent = TripShutdown
 				.newBuilder()
 				.setHeader(header)
-				.setTripDistanceMeters(mileage)
-				.setTripDurationSeconds(utcTime)
+				.setTripDistanceMeters((int) mileage)
+				.setTripDurationSeconds((int) utcTime)
 				.setTripId(tripIdentifier)
 				.build();
